# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.1.0
  description: Amrita Connect university networking API
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /auth/register:
    post:
      operationId: register
      tags: [auth]
      summary: Register a new university member
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/RegisterInput"
      responses:
        "201":
          description: Account created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AuthResponse"
        "400":
          $ref: "#/components/responses/BadRequest"
        "409":
          description: Duplicate email
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "403":
          description: Role requires administrator provisioning
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
  /auth/login:
    post:
      operationId: login
      tags: [auth]
      summary: Log in to Amrita Connect
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/LoginInput"
      responses:
        "200":
          description: Authenticated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AuthResponse"
        "401":
          description: Invalid credentials
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
  /auth/me:
    get:
      operationId: getCurrentUser
      tags: [auth]
      security:
        - bearerAuth: []
      summary: Get the current authenticated user
      responses:
        "200":
          description: Current user
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/User"
        "401":
          $ref: "#/components/responses/Unauthorized"
  /users/me:
    patch:
      operationId: updateMyProfile
      tags: [users]
      security:
        - bearerAuth: []
      summary: Update the current user's profile
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ProfileUpdate"
      responses:
        "200":
          description: Updated profile
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/User"
        "400":
          $ref: "#/components/responses/BadRequest"
        "401":
          $ref: "#/components/responses/Unauthorized"
  /users:
    get:
      operationId: listUsers
      tags: [users]
      security:
        - bearerAuth: []
      summary: Search and filter university members
      parameters:
        - $ref: "#/components/parameters/Search"
        - $ref: "#/components/parameters/Role"
        - $ref: "#/components/parameters/Campus"
        - $ref: "#/components/parameters/Department"
        - $ref: "#/components/parameters/Page"
        - $ref: "#/components/parameters/PageSize"
      responses:
        "200":
          description: Paginated people results
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/UserListResponse"
        "401":
          $ref: "#/components/responses/Unauthorized"
  /users/{id}:
    get:
      operationId: getUser
      tags: [users]
      security:
        - bearerAuth: []
      summary: View a member profile
      parameters:
        - $ref: "#/components/parameters/UserId"
      responses:
        "200":
          description: Member profile
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/PublicUser"
        "404":
          $ref: "#/components/responses/NotFound"
  /dashboard/summary:
    get:
      operationId: getDashboardSummary
      tags: [dashboard]
      security:
        - bearerAuth: []
      summary: Get a role-aware dashboard summary
      responses:
        "200":
          description: Dashboard summary
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/DashboardSummary"
        "401":
          $ref: "#/components/responses/Unauthorized"
  /mentorship/requests:
    get:
      operationId: listMentorshipRequests
      tags: [mentorship]
      security:
        - bearerAuth: []
      summary: List mentorship requests for the current user
      responses:
        "200":
          description: Mentorship requests
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/MentorshipRequest"
    post:
      operationId: createMentorshipRequest
      tags: [mentorship]
      security:
        - bearerAuth: []
      summary: Request mentorship from another member
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/MentorshipRequestInput"
      responses:
        "201":
          description: Request created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/MentorshipRequest"
        "400":
          $ref: "#/components/responses/BadRequest"
  /mentorship/requests/{id}/status:
    patch:
      operationId: updateMentorshipRequestStatus
      tags: [mentorship]
      security:
        - bearerAuth: []
      summary: Accept or reject a mentorship request
      parameters:
        - $ref: "#/components/parameters/RequestId"
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/StatusUpdate"
      responses:
        "200":
          description: Updated request
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/MentorshipRequest"
  /collaborations:
    get:
      operationId: listCollaborations
      tags: [collaborations]
      security:
        - bearerAuth: []
      summary: List collaboration opportunities
      parameters:
        - $ref: "#/components/parameters/Search"
        - name: category
          in: query
          schema:
            type: string
        - $ref: "#/components/parameters/Page"
        - $ref: "#/components/parameters/PageSize"
      responses:
        "200":
          description: Collaboration posts
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/CollaborationListResponse"
    post:
      operationId: createCollaboration
      tags: [collaborations]
      security:
        - bearerAuth: []
      summary: Create a collaboration opportunity
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CollaborationInput"
      responses:
        "201":
          description: Collaboration created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Collaboration"
  /opportunities:
    get:
      operationId: listOpportunities
      tags: [opportunities]
      security:
        - bearerAuth: []
      summary: List career and research opportunities
      parameters:
        - $ref: "#/components/parameters/Search"
        - name: category
          in: query
          schema:
            type: string
        - $ref: "#/components/parameters/Page"
        - $ref: "#/components/parameters/PageSize"
      responses:
        "200":
          description: Opportunities
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/OpportunityListResponse"
  /events:
    get:
      operationId: listEvents
      tags: [events]
      security:
        - bearerAuth: []
      summary: List upcoming university events
      parameters:
        - $ref: "#/components/parameters/Campus"
        - $ref: "#/components/parameters/Page"
        - $ref: "#/components/parameters/PageSize"
      responses:
        "200":
          description: Upcoming events
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/EventListResponse"
  /notifications:
    get:
      operationId: listNotifications
      tags: [notifications]
      security:
        - bearerAuth: []
      summary: List notifications
      responses:
        "200":
          description: Notifications
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Notification"
  /notifications/{id}/read:
    patch:
      operationId: markNotificationRead
      tags: [notifications]
      security:
        - bearerAuth: []
      summary: Mark a notification as read
      parameters:
        - $ref: "#/components/parameters/NotificationId"
      responses:
        "200":
          description: Updated notification
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Notification"
  /admin/summary:
    get:
      operationId: getAdminSummary
      tags: [admin]
      security:
        - bearerAuth: []
      summary: Get platform summary metrics for administrators
      responses:
        "200":
          description: Platform summary
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AdminSummary"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "403":
          description: Administrator access required
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  parameters:
    UserId:
      name: id
      in: path
      required: true
      schema:
        type: string
    RequestId:
      name: id
      in: path
      required: true
      schema:
        type: string
...
    Search:
      name: search
      in: query
      schema:
        type: string
    Role:
      name: role
      in: query
      schema:
        $ref: "#/components/schemas/UserRole"
    Campus:
      name: campus
      in: query
      schema:
        type: string
    Department:
      name: department
      in: query
      schema:
        type: string
    Page:
      name: page
      in: query
      schema:
        type: integer
        minimum: 1
        default: 1
    PageSize:
      name: pageSize
      in: query
      schema:
        type: integer
        minimum: 1
        maximum: 50
        default: 12
  responses:
    BadRequest:
      description: Invalid request
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/ErrorResponse"
    Unauthorized:
      description: Authentication required
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/ErrorResponse"
    NotFound:
      description: Resource not found
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/ErrorResponse"
  schemas:
...
    ErrorResponse:
      type: object
      properties:
        success:
          type: boolean
        message:
          type: string
      required: [success, message]
...
    User:
      type: object
      properties:
        id:
          type: string
        fullName:
          type: string
        email:
          type: string
        role:
          $ref: "#/components/schemas/UserRole"
        campus:
          type: string
        department:
          type: string
        graduationYear:
          type: ["integer", "null"]
        headline:
          type: string
        bio:
          type: string
        company:
          type: ["string", "null"]
        jobRole:
          type: ["string", "null"]
        skills:
          type: array
          items:
            type: string
        interests:
          type: array
          items:
            type: string
        helpWith:
          type: array
          items:
            type: string
        lookingFor:
          type: array
          items:
            type: string
        avatarUrl:
          type: ["string", "null"]
        verified:
          type: boolean
        createdAt:
          type: string
          format: date-time
      required:
        - id
        - fullName
        - email
        - role
        - campus
        - department
        - graduationYear
        - headline
        - bio
        - company
        - jobRole
        - skills
        - interests
        - helpWith
        - lookingFor
        - avatarUrl
        - verified
        - createdAt
    PublicUser:
      type: object
      properties:
        id:
          type: string
        fullName:
          type: string
        role:
          $ref: "#/components/schemas/UserRole"
        campus:
          type: string
        department:
          type: string
        graduationYear:
          type: ["integer", "null"]
        headline:
          type: string
        bio:
          type: string
        company:
          type: ["string", "null"]
        jobRole:
          type: ["string", "null"]
        skills:
          type: array
          items:
            type: string
        interests:
          type: array
          items:
            type: string
        helpWith:
          type: array
          items:
            type: string
        lookingFor:
          type: array
          items:
            type: string
        avatarUrl:
          type: ["string", "null"]
        verified:
          type: boolean
        createdAt:
          type: string
          format: date-time
      required:
        - id
        - fullName
        - role
        - campus
        - department
        - graduationYear
        - headline
        - bio
        - company
        - jobRole
        - skills
        - interests
        - helpWith
        - lookingFor
        - avatarUrl
        - verified
        - createdAt
    RegisterInput:
      type: object
      properties:
        fullName:
          type: string
          minLength: 2
        email:
          type: string
          format: email
        password:
          type: string
          minLength: 8
        role:
          $ref: "#/components/schemas/RegistrationRole"
        campus:
          type: string
        department:
          type: string
        graduationYear:
          type: ["integer", "null"]
      required: [fullName, email, password, role, campus, department]
    LoginInput:
      type: object
      properties:
        email:
          type: string
          format: email
        password:
          type: string
      required: [email, password]
    AuthResponse:
      type: object
      properties:
        token:
          type: string
        user:
          $ref: "#/components/schemas/User"
      required: [token, user]
    ProfileUpdate:
      type: object
      properties:
        headline:
          type: string
        bio:
          type: string
        company:
          type: ["string", "null"]
        jobRole:
          type: ["string", "null"]
        skills:
          type: array
          items:
            type: string
        interests:
          type: array
          items:
            type: string
        helpWith:
          type: array
          items:
            type: string
        lookingFor:
          type: array
          items:
            type: string
        avatarUrl:
          type: ["string", "null"]
    UserListResponse:
      type: object
      properties:
        items:
          type: array
          items:
            $ref: "#/components/schemas/PublicUser"
        page:
          type: integer
        pageSize:
          type: integer
        total:
          type: integer
      required: [items, page, pageSize, total]
    DashboardSummary:
      type: object
      properties:
        profileCompletion:
          type: integer
        peopleCount:
          type: integer
        mentorshipPending:
          type: integer
        savedOpportunities:
          type: integer
        upcomingEvents:
          type: integer
        unreadNotifications:
          type: integer
        recentPeople:
          type: array
          items:
            $ref: "#/components/schemas/PublicUser"
        upcoming:
          type: array
          items:
            $ref: "#/components/schemas/Event"
        opportunities:
          type: array
          items:
            $ref: "#/components/schemas/Opportunity"
      required:
        - profileCompletion
        - peopleCount
        - mentorshipPending
        - savedOpportunities
        - upcomingEvents
        - unreadNotifications
        - recentPeople
        - upcoming
        - opportunities
...
    MentorshipRequestInput:
      type: object
      properties:
        mentorId:
          type: string
        message:
          type: string
        reason:
          type: string
        topic:
          type: string
      required: [mentorId, message, reason, topic]
    StatusUpdate:
      type: object
      properties:
        status:
          type: string
          enum: [accepted, rejected]
      required: [status]
    MentorshipRequest:
      type: object
      properties:
        id:
          type: string
        mentor:
          $ref: "#/components/schemas/PublicUser"
        requester:
          $ref: "#/components/schemas/PublicUser"
        message:
          type: string
        reason:
          type: string
        topic:
          type: string
        status:
          type: string
          enum: [pending, accepted, rejected]
        createdAt:
          type: string
          format: date-time
      required: [id, mentor, requester, message, reason, topic, status, createdAt]
...
    CollaborationListResponse:
      type: object
      properties:
        items:
          type: array
          items:
            $ref: "#/components/schemas/Collaboration"
        page:
          type: integer
        pageSize:
          type: integer
        total:
          type: integer
      required: [items, page, pageSize, total]
...